magicmind/tools/preprocess/requirements.txt
numpy >= 1.19.5
scipy == 1.5.4
scikit-build == 0.11.1
opencv-python == 4.5.1.48
pyyaml
tqdm
